requirements.txt
# Core
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
SQLAlchemy>=1.4.0
pandas>=1.3.0
openpyxl>=3.0.7
tqdm>=4.62.0
ijson>=3.1.4
orjson>=3.6.4
chardet>=4.0.0

# Web Framework (for future API)
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Development
pytest>=6.2.5
black>=21.9b0
isort>=5.9.3
